+++ ReleaseNotes.md
@@ -4,12 +4,15 @@
 - [JCTools.GenericCrud Release notes](#jctoolsgenericcrud-release-notes)
   - [Content](#content)
   - [Version 2.2.0](#version-220)
+    - [beta3](#beta3)
     - [beta2](#beta2)
     - [beta1](#beta1)
   - [Version 2.1.0](#version-210)
   - [Version 2.0.0](#version-200)
 
 ## Version 2.2.0
+### beta3 
+* Support to delete of entities by API Rest added.  
 ### beta2
 * Support to get the entities details by API Rest added.
 ### beta1